Бонусы и финальный рефакторинг

+++ src/main/java/school21/smartcalc/models/CreditRequest.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class CreditRequest {
     private final double amount;
-    private final int period;
+    private final double period;
     private final double percent;
     private final String type;
 }
